Replace kybra_like to project-neutral example
